Split exercise card system tests
